Separate wasm and IBC modules registration
This commit separates the wasm and IBC modules registration to enhance code clarity and organization. The registration of wasm modules is now done in app/wasm.go as opposed to being coupled with IBC registration in app/ibc.go. The changes make the code more modular, which can make future updates or modifications easier and safer to implement.